+++ crates/starknet-server/src/api/mod.rs
@@ -5,7 +5,7 @@ pub(crate) mod serde_helpers;
 
 use std::sync::Arc;
 
-use starknet_core::Starknet;
+use starknet_core::starknet::Starknet;
 use tokio::sync::RwLock;
 
 /// Data that can be shared between threads with read write lock access
