--- crates/starknet/src/services/mod.rs
@@ -1,2 +1,0 @@
-mod add_declare_transaction;
-mod predeployed;
