+++ crates/starknet/src/starknet/add_declare_transaction.rs
@@ -3,140 +3,139 @@ use starknet_types::felt::{ClassHash, TransactionHash};
 use starknet_types::traits::HashProducer;
 
 use crate::error::{Error, Result};
+use crate::starknet::Starknet;
 use crate::transactions::declare_transaction::DeclareTransactionV1;
 use crate::transactions::declare_transaction_v2::DeclareTransactionV2;
 use crate::transactions::{StarknetTransaction, Transaction};
-use crate::Starknet;
 
-impl Starknet {
-    pub fn add_declare_transaction_v2(
-        &mut self,
-        declare_transaction: DeclareTransactionV2,
-    ) -> Result<(TransactionHash, ClassHash)> {
-        let mut declare_transaction = declare_transaction;
+pub fn add_declare_transaction_v2(
+    starknet: &mut Starknet,
+    declare_transaction: DeclareTransactionV2,
+) -> Result<(TransactionHash, ClassHash)> {
+    let mut declare_transaction = declare_transaction;
 
-        let class_hash = declare_transaction.sierra_contract_class.generate_hash()?;
-        declare_transaction.class_hash = Some(class_hash);
+    let class_hash = declare_transaction.sierra_contract_class.generate_hash()?;
+    declare_transaction.class_hash = Some(class_hash);
 
-        let transaction_hash = declare_transaction.generate_hash()?;
-        declare_transaction.transaction_hash = Some(transaction_hash);
+    let transaction_hash = declare_transaction.generate_hash()?;
+    declare_transaction.transaction_hash = Some(transaction_hash);
 
-        // TODO compile the contract and check the hash
-        let transaction = DeclareV2::new(
-            &declare_transaction.sierra_contract_class.clone().try_into()?,
-            None,
-            declare_transaction.compiled_class_hash.into(),
-            declare_transaction.chain_id.into(),
-            declare_transaction.sender_address.try_into()?,
-            declare_transaction.max_fee,
-            declare_transaction.version().into(),
-            declare_transaction.signature.iter().map(|felt| felt.into()).collect(),
-            declare_transaction.nonce.into(),
-        )?;
+    // TODO compile the contract and check the hash
+    let transaction = DeclareV2::new(
+        &declare_transaction.sierra_contract_class.clone().try_into()?,
+        None,
+        declare_transaction.compiled_class_hash.into(),
+        declare_transaction.chain_id.into(),
+        declare_transaction.sender_address.try_into()?,
+        declare_transaction.max_fee,
+        declare_transaction.version().into(),
+        declare_transaction.signature.iter().map(|felt| felt.into()).collect(),
+        declare_transaction.nonce.into(),
+    )?;
 
-        if transaction.max_fee == 0 {
-            return Err(Error::TransactionError(
-                starknet_in_rust::transaction::error::TransactionError::FeeError(
-                    "For declare transaction version 2, max fee cannot be 0".to_string(),
-                ),
-            ));
-        }
-
-        let state_before_txn = self.state.pending_state.clone();
-
-        match transaction.execute(&mut self.state.pending_state, &self.block_context) {
-            Ok(tx_info) => {
-                declare_transaction.class_hash = Some(class_hash);
-
-                self.handle_successful_transaction(
-                    &transaction_hash,
-                    Transaction::DeclareV2(declare_transaction),
-                    tx_info,
-                )?;
-            }
-            Err(tx_err) => {
-                let transaction_to_add = StarknetTransaction::create_rejected(
-                    Transaction::DeclareV2(declare_transaction),
-                    tx_err,
-                );
-
-                self.transactions.insert(&transaction_hash, transaction_to_add);
-                // Revert to previous pending state
-                self.state.pending_state = state_before_txn;
-            }
-        }
-
-        Ok((transaction_hash, class_hash))
+    if transaction.max_fee == 0 {
+        return Err(Error::TransactionError(
+            starknet_in_rust::transaction::error::TransactionError::FeeError(
+                "For declare transaction version 2, max fee cannot be 0".to_string(),
+            ),
+        ));
     }
-    pub fn add_declare_transaction_v1(
-        &mut self,
-        declare_transaction: DeclareTransactionV1,
-    ) -> Result<(TransactionHash, ClassHash)> {
-        let mut declare_transaction = declare_transaction;
 
-        let class_hash = declare_transaction.contract_class.generate_hash()?;
-        let transaction_hash = declare_transaction.generate_hash()?;
-        declare_transaction.transaction_hash = Some(transaction_hash);
+    let state_before_txn = starknet.state.pending_state.clone();
 
-        let transaction = Declare {
-            class_hash: class_hash.into(),
-            sender_address: declare_transaction.sender_address.try_into()?,
-            tx_type: starknet_in_rust::definitions::transaction_type::TransactionType::Declare,
-            validate_entry_point_selector:
-                starknet_in_rust::definitions::constants::VALIDATE_DECLARE_ENTRY_POINT_SELECTOR
-                    .clone(),
-            version: declare_transaction.version().into(),
-            max_fee: declare_transaction.max_fee,
-            signature: declare_transaction.signature.iter().map(|felt| felt.into()).collect(),
-            nonce: declare_transaction.nonce.into(),
-            hash_value: transaction_hash.into(),
-            contract_class: declare_transaction.contract_class.clone().try_into()?,
-            skip_execute: false,
-            skip_fee_transfer: false,
-            skip_validate: false,
-        };
+    match transaction.execute(&mut starknet.state.pending_state, &starknet.block_context) {
+        Ok(tx_info) => {
+            declare_transaction.class_hash = Some(class_hash);
 
-        verify_version(
-            &transaction.version,
-            transaction.max_fee,
-            &transaction.nonce,
-            &transaction.signature,
-        )?;
-
-        if transaction.max_fee == 0 {
-            return Err(Error::TransactionError(
-                starknet_in_rust::transaction::error::TransactionError::FeeError(
-                    "For declare transaction version 1, max fee cannot be 0".to_string(),
-                ),
-            ));
+            // Add sierra contract
+            starknet.sierra_contracts.insert(class_hash, transaction.sierra_contract_class);
+            starknet.handle_successful_transaction(
+                &transaction_hash,
+                Transaction::DeclareV2(declare_transaction),
+                tx_info,
+            )?;
         }
+        Err(tx_err) => {
+            let transaction_to_add = StarknetTransaction::create_rejected(
+                Transaction::DeclareV2(declare_transaction),
+                tx_err,
+            );
 
-        let state_before_txn = self.state.pending_state.clone();
-
-        match transaction.execute(&mut self.state.pending_state, &self.block_context) {
-            Ok(tx_info) => {
-                declare_transaction.class_hash = Some(class_hash);
-
-                self.handle_successful_transaction(
-                    &transaction_hash,
-                    Transaction::Declare(declare_transaction),
-                    tx_info,
-                )?;
-            }
-            Err(tx_err) => {
-                let transaction_to_add = StarknetTransaction::create_rejected(
-                    Transaction::Declare(declare_transaction),
-                    tx_err,
-                );
-
-                self.transactions.insert(&transaction_hash, transaction_to_add);
-                // Revert to previous pending state
-                self.state.pending_state = state_before_txn;
-            }
+            starknet.transactions.insert(&transaction_hash, transaction_to_add);
+            // Revert to previous pending state
+            starknet.state.pending_state = state_before_txn;
         }
-
-        Ok((transaction_hash, class_hash))
     }
+
+    Ok((transaction_hash, class_hash))
+}
+pub fn add_declare_transaction_v1(
+    starknet: &mut Starknet,
+    declare_transaction: DeclareTransactionV1,
+) -> Result<(TransactionHash, ClassHash)> {
+    let mut declare_transaction = declare_transaction;
+
+    let class_hash = declare_transaction.contract_class.generate_hash()?;
+    let transaction_hash = declare_transaction.generate_hash()?;
+    declare_transaction.transaction_hash = Some(transaction_hash);
+
+    let transaction = Declare {
+        class_hash: class_hash.into(),
+        sender_address: declare_transaction.sender_address.try_into()?,
+        tx_type: starknet_in_rust::definitions::transaction_type::TransactionType::Declare,
+        validate_entry_point_selector:
+            starknet_in_rust::definitions::constants::VALIDATE_DECLARE_ENTRY_POINT_SELECTOR.clone(),
+        version: declare_transaction.version().into(),
+        max_fee: declare_transaction.max_fee,
+        signature: declare_transaction.signature.iter().map(|felt| felt.into()).collect(),
+        nonce: declare_transaction.nonce.into(),
+        hash_value: transaction_hash.into(),
+        contract_class: declare_transaction.contract_class.clone().try_into()?,
+        skip_execute: false,
+        skip_fee_transfer: false,
+        skip_validate: false,
+    };
+
+    verify_version(
+        &transaction.version,
+        transaction.max_fee,
+        &transaction.nonce,
+        &transaction.signature,
+    )?;
+
+    if transaction.max_fee == 0 {
+        return Err(Error::TransactionError(
+            starknet_in_rust::transaction::error::TransactionError::FeeError(
+                "For declare transaction version 1, max fee cannot be 0".to_string(),
+            ),
+        ));
+    }
+
+    let state_before_txn = starknet.state.pending_state.clone();
+
+    match transaction.execute(&mut starknet.state.pending_state, &starknet.block_context) {
+        Ok(tx_info) => {
+            declare_transaction.class_hash = Some(class_hash);
+
+            starknet.handle_successful_transaction(
+                &transaction_hash,
+                Transaction::Declare(declare_transaction),
+                tx_info,
+            )?;
+        }
+        Err(tx_err) => {
+            let transaction_to_add = StarknetTransaction::create_rejected(
+                Transaction::Declare(declare_transaction),
+                tx_err,
+            );
+
+            starknet.transactions.insert(&transaction_hash, transaction_to_add);
+            // Revert to previous pending state
+            starknet.state.pending_state = state_before_txn;
+        }
+    }
+
+    Ok((transaction_hash, class_hash))
 }
 
 #[cfg(test)]
@@ -155,12 +154,12 @@ mod tests {
     use crate::account::Account;
     use crate::constants::{self};
     use crate::error::Error;
+    use crate::starknet::{predeployed, Starknet};
     use crate::traits::{Accounted, HashIdentifiedMut, StateChanger};
     use crate::transactions::declare_transaction::DeclareTransactionV1;
     use crate::transactions::declare_transaction_v2::DeclareTransactionV2;
     use crate::utils::load_cairo_0_contract_class;
     use crate::utils::test_utils::dummy_felt;
-    use crate::Starknet;
 
     fn test_declare_transaction_v2(sender_address: ContractAddress) -> DeclareTransactionV2 {
         let contract_json_path = concat!(
@@ -229,7 +228,7 @@ mod tests {
         let initial_cached_state =
             starknet.state.pending_state.casm_contract_classes().as_ref().unwrap().len();
         let declare_txn = test_declare_transaction_v2(sender);
-        let (txn_hash, _) = starknet.add_declare_transaction_v2(declare_txn).unwrap();
+        let (txn_hash, class_hash) = starknet.add_declare_transaction_v2(declare_txn).unwrap();
         let txn = starknet.transactions.get_by_hash_mut(&txn_hash).unwrap();
 
         assert_eq!(txn.status, TransactionStatus::Rejected);
@@ -237,6 +236,7 @@ mod tests {
             initial_cached_state,
             starknet.state.pending_state.casm_contract_classes().as_ref().unwrap().len()
         );
+        assert!(starknet.sierra_contracts.get(&class_hash).is_none())
     }
 
     #[test]
@@ -252,6 +252,7 @@ mod tests {
         assert_eq!(class_hash, declare_txn.sierra_contract_class.generate_hash().unwrap());
         // check if txn is with status accepted
         assert_eq!(tx.status, TransactionStatus::AcceptedOnL2);
+        assert!(starknet.sierra_contracts.get(&class_hash).is_some())
     }
 
     #[test]
@@ -332,7 +333,7 @@ mod tests {
         );
         let contract_class = load_cairo_0_contract_class(account_json_path).unwrap();
 
-        let erc_20_contract = Starknet::create_erc20().unwrap();
+        let erc_20_contract = predeployed::create_erc20().unwrap();
         erc_20_contract.deploy(&mut starknet.state).unwrap();
 
         let acc = Account::new(
